refactor: move arguments types to types.arguments
